Added vaults section to address
